feat(config): Global config (#221)
Update config data, now besides a project's local config file we have a
global config file as well. If some value like chain-id is not set on
either, we have a default value we use. This allows queries to be run on
any directory, even outside of a CLI project.

New Commands to set config directly:

> archway config chain-id
> archway config contracts-path
> archway config keyring-backend
> archway config keyring-path
> archway config default-account

Closes #213