fix(plugins): Improvements to the help output (#253)
- Help command now display nested subcommand help info e.g. `archway
help contracts new`
- Verified the default and required values being displayed.
- Sorted required arguments at the start of arguments list. Separated
required flags from the other flags, also separated transaction and
keyring flags in a separate group.
- Verified command examples 
- Removed index files on nested subcommands because it caused the help
module to treat them as regular commands. The descriptions for them are
added on package.json. Only downside is that this makes them not
available for unit tests, those were removed

Closes #222

---------

Co-authored-by: Augusto Elesbão <aelesbao@users.noreply.github.com>